% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_api_info.R
\name{ya_api_info}
\alias{ya_api_info}
\title{Title: ya_api_info()}
\usage{
ya_api_info(lang = "EN")
}
\arguments{
\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}
}
\value{
A string with type, status code and description, and api version in Portuguese an English languages, and a generalized error message to check internet connection.
}
\description{
Description: Shows useful informations such as type, status code and description, and version of current BPstat Data API.
}
\examples{
#'ya_api_info()
}
