% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_get_dataset.R
\name{ya_get_dataset}
\alias{ya_get_dataset}
\title{Title: ya_get_dataset()}
\usage{
ya_get_dataset(lang = "EN")
}
\arguments{
\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}
}
\value{
A data table with the information about the BPstat domains
}
\description{
Description: Gathers information about the datasets in BPstat.
}
\examples{
#'ya_get_dataset()
}
