% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_get_dataset_item.R
\name{ya_get_dataset_item}
\alias{ya_get_dataset_item}
\title{Title: ya_get_dataset_item()}
\usage{
ya_get_dataset_item(lang = "EN")
}
\arguments{
\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}
}
\value{
A data table with the information in a more human readable form with a search option using regex.
}
\description{
Description: A function that gathers information available in BPstat about the dataset items.
}
\examples{
ya_get_dataset_item()
}
