% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_get_dimension.R
\name{ya_get_dimension}
\alias{ya_get_dimension}
\title{Title: ya_get_dimension()}
\usage{
ya_get_dimension(lang = "EN")
}
\arguments{
\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}
}
\value{
A DT datatable in RStudio viewer pane
}
\description{
Description: A function gathers information about BPstat domains located at "https://bpstat.bportugal.pt/data/v1/domains/?lang=" ("PT" or "EN")
}
\examples{
#'\dontrun{
#'ya_get_dimension()
#'}
}
