% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_get_dimension_description.R
\name{ya_get_dimension_description}
\alias{ya_get_dimension_description}
\title{Title ya_get_dimension_description()}
\usage{
ya_get_dimension_description(lang = "EN")
}
\arguments{
\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}
}
\value{
A DT datatable in RStudio viewer pane
}
\description{
Description: A function that gathers information avaliable in BPstat about the dimension items and description. This function takes a little longer to execute.
}
\examples{
#'ya_get_dimension_description()
}
