% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_get_domain.R
\name{ya_get_domain}
\alias{ya_get_domain}
\title{Title: ya_get_domain()}
\usage{
ya_get_domain(lang = "EN")
}
\arguments{
\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}
}
\value{
A DT table in the RStudio viewer
}
\description{
Description: This function gathers information about BPstat domains and makes visible to the user in a table format for better visualization and comprehension. The returned fields are:
ID - The unique ID of the domain in the BPstat structure and identification.
Parent ID - The ID of the parent section (none if it is a top level domain)
Order - The order where it is displayed in the BPstat theme organization
Label - The label of the domain
Short label - Short label of the domain
Description - The domain description
Has series - If the domain has series (True if it has, False if it has not)
Last update - Last date when the domain has been updated
Number of series - Number of series that the domain has
Number of datasets - Number of datasets that the domain has
Domain URL - The Uniform Resource Location of the domain in BPstat
Dataset URL - The Uniform Resource Location of the dataset in BPstat
Dimension URL - The Uniform Resource Location of the dimension in BPstat
}
\examples{
#'ya_get_domain()
}
