% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_info_serie.R
\name{ya_info_serie}
\alias{ya_info_serie}
\title{Title: ya_info_serie()}
\usage{
ya_info_serie(series_id = 0, operation = "P", lang = "EN")
}
\arguments{
\item{series_id}{A serie ID from BPstat. Defaults to 0 (zero)}

\item{operation}{A string with the length of 1 (one), representing the following operations:
P - plot
T - table
Df - dataframe}

\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}
}
\value{
From the user choice can be a plot, a table or a tibble for further analysis
}
\description{
Description: A function allows to user to plot, see in a table or reuse in other context, data from a determined BPstat serie ID.
}
\examples{
#'ya_info_serie(series_id = 12528985)
}
