% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_my_favourites.R
\name{ya_my_favourites}
\alias{ya_my_favourites}
\title{Title: ya_my_favourites()}
\usage{
ya_my_favourites(series_id = 0, operation = "V", lang = "EN")
}
\arguments{
\item{series_id}{A serie ID from BPstat. Defaults to 0 (zero)}

\item{operation}{A string with the length of 1 (one), representing the following operations:
A - add
D - delete
V - view}

\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}
}
\value{
A file named «ya_favourites» with user favourite series
}
\description{
Description: this function allows user to add, delete and view favourite series from BPStat
}
\examples{
ya_my_favourites()
}
