% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_save_dataset.R
\name{ya_save_dataset}
\alias{ya_save_dataset}
\title{Title ya_save_dataset()}
\usage{
ya_save_dataset(lang = "EN", extension = "rds")
}
\arguments{
\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}

\item{extension}{A string that represents the avaliable extensions to save files. Currently only uses "rds" or "csv". Defaults to "rds"}
}
\value{
A file with the information about BPstat datasets with the rds or csv extension.
}
\description{
Description A function that saves de information avaliable about the BPstat datasets
}
\examples{
ya_save_dataset()
}
