% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_save_dimension_description.R
\name{ya_save_dimension_description}
\alias{ya_save_dimension_description}
\title{Title ya_save_dimension_description()}
\usage{
ya_save_dimension_description(lang = "EN", extension = "rds")
}
\arguments{
\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}

\item{extension}{A string that represents the available extensions to save files. Currently only uses "rds" or "csv". Defaults to "rds"}
}
\value{
A file with the information about BPstat description of dimensions with the rds or csv extension.
}
\description{
Description: A function that saves de information available about the BPstat description of dimensions.
}
\examples{
#'\dontrun{
#'ya_save_dimension_description()
#'}
}
