% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_save_dimension_item.R
\name{ya_save_dimension_item}
\alias{ya_save_dimension_item}
\title{Title: ya_save_dimension_item()}
\usage{
ya_save_dimension_item(lang = "EN", extension = "rds")
}
\arguments{
\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}

\item{extension}{A string that represents the available extensions to save files. Currently only uses "rds" or "csv". Defaults to "rds"}
}
\value{
A file with the information about BPstat dimensions items with the rds or csv extension.
}
\description{
Description: A function that saves de information available about the BPstat dimensions items.
}
\examples{
#'\dontrun{
#'ya_save_dimension_item()
#'}
}
