% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_save_domain.R
\name{ya_save_domain}
\alias{ya_save_domain}
\title{Title: ya_save_domain()}
\usage{
ya_save_domain(lang = "EN", extension = "rds")
}
\arguments{
\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}

\item{extension}{A string that represents the avaliable extensions to save files. Currently only uses "rds" or "csv". Defaults to "rds"}
}
\value{
A file with the available data from the domains in the rds or csv format.
}
\description{
Description: A function that saves de information avaliable about the BPstat domains
}
\examples{
ya_save_domain()
}
