% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_search.R
\name{ya_search}
\alias{ya_search}
\title{Title: ya_search()}
\usage{
ya_search(query, lang = "EN")
}
\arguments{
\item{query}{A string that represents a query to be executed in BPstat}

\item{lang}{A string that represents the languages supported by BPstat API. Currently only uses "PT" or "EN". Defaults to "EN"}
}
\value{
A data table with the query results from BPstat.
}
\description{
Description: This is a function that allows to the user search directly de BPstat data through the API.
}
\examples{
#'\dontrun{
#'ya_search(query="ihpc")
#'}

}
